porting logic to ExternalApplication small fixes
